Fixed #110

+++ src/main/java/lorienlegacies/legacies/PlayerLegacyData.java
@@ -15,10 +15,10 @@ public class PlayerLegacyData
 	// Non-saved data
 	public LinkedHashMap<String, Boolean> legacyToggles = new LinkedHashMap<String, Boolean>();
 	public LinkedHashMap<String, Integer> legacyPrevLvl = new LinkedHashMap<String, Integer>();
-	public int stamina;
+	public float stamina;
 	
 	// The client needs to know the max amount of stamina without overwriting the config (well, as of 1.12.2, and it's done now)
-	public int maxClientStamina;
+	public float maxClientStamina;
 	
 	// The client also needs to know their legacies' levels
 	public LinkedHashMap<String, Integer> clientLegacyLevels = new LinkedHashMap<String, Integer>();
